% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{add_cols}
\alias{add_cols}
\title{Add new columns to a data frame}
\usage{
add_cols(data, cols, pos)
}
\arguments{
\item{data}{a data frame where new columns are inserted}

\item{cols}{named list of columns to add to `data`}

\item{pos}{positions to insert columns after}
}
\description{
Extends `tibble::add_column` to insert columns to multiple positions.
}
