% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_pin.R
\name{is_valid_pin}
\alias{is_valid_pin}
\title{Check if valid PIN}
\usage{
is_valid_pin(x, ...)
}
\arguments{
\item{x}{vector to check for PIN validity}
}
\value{
logical vector of `length(x)`
}
\description{
Check if valid PIN
}
