% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Create a PIN}
\usage{
pin(dob, ordinal)
}
\arguments{
\item{dob}{Date vector of dates of birth}

\item{ordinal}{integer giving order of birth on `dob`. Even for female, odd
for male. Takes values between 1 and 999.}
}
\description{
Create a PIN
}
