% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_extractors.R
\name{pin_extractors}
\alias{pin_extractors}
\alias{pin_dob}
\alias{pin_sex}
\title{Extract data from PIN}
\usage{
pin_dob(x, try_fix = FALSE)

pin_sex(x, factor = TRUE, language = c("english", "finnish"))
}
\arguments{
\item{x}{Character vector of PINs.}

\item{try_fix}{Logical. Should missing century be replaced with 19?}

\item{factor}{Logical. Should the return value be a factor?}

\item{language}{Character scalar. Language to use for factor labels if
\code{factor = TRUE}. Currently either \code{english} or \code{finnish}.}
}
\description{
Extract data from PIN
}
\section{Functions}{
\itemize{
\item \code{pin_dob}: Extract date of birth from PIN

\item \code{pin_sex}: Extract sex from PIN
}}

