% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_rescue.R
\name{pin_rescue}
\alias{pin_rescue}
\title{Attempt to fix common problems in PINs}
\usage{
pin_rescue(x, default_sep = c("-", "A", "+"), validate = FALSE)
}
\arguments{
\item{x}{character vector of PINs}

\item{default_sep}{character to use as default separator}

\item{validate}{should result be checked to ensure all PINs are valid?}
}
\description{
Currently checks if the separator is valid and inserts `default_sep` if not.
}
