% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{pseudonymize}
\alias{pseudonymize}
\alias{pseudonymise}
\title{Pseudonymize columns in data containing PINs}
\usage{
pseudonymize(data, key, ..., guess = FALSE, replace = TRUE,
  rename = !replace, quiet = FALSE)

pseudonymise(data, key, ..., guess = FALSE, replace = TRUE,
  rename = !replace, quiet = FALSE)
}
\arguments{
\item{data}{A data frame containing PINs to be pseudonymized.}

\item{key}{Named vector or data frame, used as a lookup table for pids. If
data frame, the first column is assumed to contain PINs and the second
column the corresponding pids.}

\item{...}{Manually selected columns to be pseudonymized. These are
automatically quoted and evaluated in the context of the data. Uses
`tidyselect` semantics for selection.}

\item{guess}{Logical. Attempt to automatically identify and pseudonymize
columns that contain PINs?}

\item{replace}{Logical. Should PIN columns be replaced with the pseudonymized
versions?}

\item{rename}{Logical or function. If `FALSE`, pseudonymized columns will not
be automatically renamed; if `TRUE`, they will be suffixed with `"_pid"`;
if a function, will be called on PIN column names to generate new names for
the pseudonymized columns. Manually specified new names will always be used
regardless.}

\item{quiet}{Suppress additional messages? Currently controls showing a
warning if pseudonymization results in all `NA` values in some column.}
}
\value{
A data frame where PINs have probably been linked to pids. If
  \code{replace = TRUE} values in columns guessed to have PINs have been
  replaced with matching pids from `key`.
}
\description{
Pseudonymize columns in data containing PINs
}
\seealso{
\code{\link{is_probably_pin}} used to guess if columns contain PINs
}
