% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_probably_pin.R
\name{is_probably_pin}
\alias{is_probably_pin}
\title{Guess if a vector contains PINs}
\usage{
is_probably_pin(x, ...)
}
\arguments{
\item{x}{Vector to test for pin-iness.}

\item{...}{Further arguments passed down to methods. Currently unused.}
}
\value{
Logical. Is the vector likely to contain pins?
}
\description{
Guess if a vector contains PINs
}
