% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{pseudonymize}
\alias{pseudonymize}
\alias{pseudonymise}
\title{Pseudonymize columns in data containing PINs}
\usage{
pseudonymize(data, key, ..., guess = FALSE, remove = TRUE,
  pid_suffix = "_pid")

pseudonymise(data, key, ..., guess = FALSE, remove = TRUE,
  pid_suffix = "_pid")
}
\arguments{
\item{data}{A data frame containing PINs to be pseudonymized.}

\item{key}{Named vector or data frame, used as a lookup table for pids. If
data frame, the first column is assumed to contain PINs and the second
column the corresponding pids.}

\item{...}{Manually selected columns to be pseudonymized. These are
automatically quoted and evaluated in the context of the data. Uses
`tidyselect` semantics for selection.}

\item{guess}{Logical. Attempt to automatically identify and pseudonymize
columns that contain PINs?}

\item{remove}{Logical. Should identified PIN columns be dropped?}

\item{pid_suffix}{Character scalar. Added to the end of the pin column names
for the corresponding pid columns if \code{remove = FALSE}.}
}
\value{
A data frame where PINs have probably been linked to pids. If
  \code{remove = TRUE} values in columns guessed to have PINs have been
  replaced with matching pids from `key`.
}
\description{
Pseudonymize columns in data containing PINs
}
\seealso{
\code{\link{is_probably_pin}} used to guess if columns contain PINs
}
