% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{pseudonymize}
\alias{pseudonymize}
\alias{pseudonymise}
\title{Pseudonymize columns in data containing PINs}
\usage{
pseudonymize(data, key, ..., drop_pin = TRUE, pid_suffix = "_pid",
  guess = FALSE)

pseudonymise(data, key, ..., drop_pin = TRUE, pid_suffix = "_pid",
  guess = FALSE)
}
\arguments{
\item{data}{A data frame containing PINs to be pseudonymized.}

\item{key}{Named vector or data frame, used as a lookup table for pids. If
data frame, the first column is assumed to contain PINs and the second
column the corresponding pids.}

\item{drop_pin}{Logical. Should identified PIN columns be dropped?}

\item{pid_suffix}{Character scalar. Added to the end of the pin column names
for the corresponding pid columns if \code{drop_pin = FALSE}.}

\item{guess}{Logical. Attempt to automatically identify and pseudonymize
columns that contain PINs?}
}
\value{
A data frame where PINs have probably been linked to pids. If
  \code{drop_pin = TRUE} values in columns guessed to have PINs have been
  replaced with matching pids from `key`.
}
\description{
Pseudonymize columns in data containing PINs
}
\seealso{
\code{\link{is_probably_pin}} used to guess if columns contain PINs
}
